# internal/web/handler/auth.go
package handler

import (
	"encoding/json"
	"net/http"

	"github.com/m-bromo/go-auth-template/internal/domain"
	"github.com/m-bromo/go-auth-template/internal/service"
	clienterrors "github.com/m-bromo/go-auth-template/internal/web/client_errors"
	"github.com/m-bromo/go-auth-template/internal/web/cookie"
	"github.com/m-bromo/go-auth-template/internal/web/models"
	"github.com/m-bromo/go-auth-template/pkg/validation"
)

type AuthHandler struct {
	authService         service.AuthService
	refreshTokenService service.RefreshTokenService
	otpService          service.OtpService
	cookieManager       *cookie.CookieManager
}

func NewAuthHandler(
	authService service.AuthService,
	refreshTokenService service.RefreshTokenService,
	otpService service.OtpService,
	cookieManager *cookie.CookieManager,
) *AuthHandler {
	return &AuthHandler{
		authService:         authService,
		refreshTokenService: refreshTokenService,
		cookieManager:       cookieManager,
		otpService:          otpService,
	}
}

func (h *AuthHandler) RegisterUser(w http.ResponseWriter, r *http.Request) {
	var payload models.RegisterUserPayload
	if err := json.NewDecoder(r.Body).Decode(&payload); err != nil {
		HandleError(w, clienterrors.NewBadRequestError("invalid request body", err))
		return
	}

	if err := validation.Validator.Struct(payload); err != nil {
		HandleError(w, err)
		return
	}

	if err := h.authService.RegisterUser(r.Context(), &domain.User{
		Email:    payload.Email,
		Password: payload.Password,
		Username: payload.Username,
	}); err != nil {
		HandleError(w, err)
		return
	}

	w.WriteHeader(http.StatusCreated)
}

func (h *AuthHandler) Login(w http.ResponseWriter, r *http.Request) {
	var payload models.LoginPayload
	if err := json.NewDecoder(r.Body).Decode(&payload); err != nil {
		HandleError(w, clienterrors.NewBadRequestError("invalid request body", err))
		return
	}

	if err := validation.Validator.Struct(payload); err != nil {
		HandleError(w, err)
		return
	}

	accessToken, refreshToken, err := h.authService.Login(r.Context(), &domain.User{
		Email:    payload.Email,
		Password: payload.Password,
	})
	if err != nil {
		HandleError(w, err)
		return
	}
	h.cookieManager.SetCookie(w, refreshToken)

	response := &models.LoginResponse{
		AccessToken: accessToken,
	}

	HandleJSON(w, http.StatusOK, response)
}

func (h *AuthHandler) SendOTP(w http.ResponseWriter, r *http.Request) {
	var payload models.SendOTPPayload
	if err := json.NewDecoder(r.Body).Decode(&payload); err != nil {
		HandleError(w, clienterrors.NewBadRequestError("invalid request body", err))
		return
	}

	if err := validation.Validator.Struct(payload); err != nil {
		HandleError(w, err)
		return
	}

	challengeID, err := h.otpService.SendCode(r.Context(), payload.Email)
	if err != nil {
		HandleError(w, err)
		return
	}

	response := models.SendOTPResponse{
		ChallengeID: challengeID,
	}

	HandleJSON(w, http.StatusOK, response)
}

func (h *AuthHandler) LoginWithOtp(w http.ResponseWriter, r *http.Request) {
	var payload models.VerifyOTPPayload
	if err := json.NewDecoder(r.Body).Decode(&payload); err != nil {
		HandleError(w, clienterrors.NewBadRequestError("invalid request body", err))
		return
	}

	if err := validation.Validator.Struct(payload); err != nil {
		HandleError(w, err)
		return
	}

	accessToken, refreshToken, err := h.authService.LoginWithOtp(r.Context(), payload.Email, payload.Code)
	if err != nil {
		HandleError(w, err)
		return
	}
	h.cookieManager.SetCookie(w, refreshToken)

	response := &models.LoginResponse{
		AccessToken: accessToken,
	}

	HandleJSON(w, http.StatusOK, response)
}

func (h *AuthHandler) Refresh(w http.ResponseWriter, r *http.Request) {
	c, err := h.cookieManager.GetCookie(r)
	if err != nil {
		HandleError(w, domain.NewUnauthenticatedError("refresh token was not provided", service.ErrInvalidRefreshToken))
		return
	}

	newAccessToken, newRefreshToken, err := h.refreshTokenService.Refresh(r.Context(), c.Value)
	if err != nil {
		HandleError(w, err)
		return
	}

	h.cookieManager.SetCookie(w, newRefreshToken)

	HandleJSON(w, http.StatusOK, newAccessToken)
}

func (h *AuthHandler) Logout(w http.ResponseWriter, r *http.Request) {
	cookie, err := h.cookieManager.GetCookie(r)
	if err != nil {
		h.cookieManager.DeleteCookie(w)
		w.WriteHeader(http.StatusNoContent)
		return
	}

	if err := h.refreshTokenService.Revoke(r.Context(), cookie.Value); err != nil {
		h.cookieManager.DeleteCookie(w)
		HandleError(w, err)
		return
	}

	h.cookieManager.DeleteCookie(w)

	w.WriteHeader(http.StatusNoContent)
}

func (h *AuthHandler) VerifyPasswordResetCode(w http.ResponseWriter, r *http.Request) {
	var payload models.VerifyPasswordResetCodePayload
	if err := json.NewDecoder(r.Body).Decode(&payload); err != nil {
		HandleError(w, clienterrors.NewBadRequestError("invalid request body", err))
		return
	}

	if err := validation.Validator.Struct(payload); err != nil {
		HandleError(w, err)
		return
	}

	resetToken, err := h.otpService.VerifyPasswordResetCode(r.Context(), payload.Code, payload.ChallengeID)
	if err != nil {
		HandleError(w, err)
		return
	}

	response := models.VerifyOTPResponse{
		ResetToken: resetToken,
	}

	HandleJSON(w, http.StatusOK, response)
}

func (h *AuthHandler) ResetPassword(w http.ResponseWriter, r *http.Request) {
	var payload models.ResetPasswordPayload
	if err := json.NewDecoder(r.Body).Decode(&payload); err != nil {
		HandleError(w, clienterrors.NewBadRequestError("invalid request body", err))
		return
	}

	if err := validation.Validator.Struct(payload); err != nil {
		HandleError(w, err)
		return
	}

	err := h.authService.ResetPassword(r.Context(), payload.ResetToken, payload.Password)
	if err != nil {
		HandleError(w, err)
		return
	}

	w.WriteHeader(http.StatusNoContent)
}